top_k performance improvement
