+++ crates/polars-core/src/series/implementations/datetime.rs
@@ -290,6 +290,14 @@ impl SeriesTrait for SeriesWrap<DatetimeChunked> {
         self.0.get_any_value_unchecked(index)
     }
 
+    fn top_k(&self, k: usize, descending: bool) -> PolarsResult<Series> {
+        Ok(self
+            .0
+            .top_k(k, descending)
+            .into_datetime(self.0.time_unit(), self.0.time_zone().clone())
+            .into_series())
+    }
+
     fn sort_with(&self, options: SortOptions) -> PolarsResult<Series> {
         Ok(self
             .0
