+++ crates/polars-core/src/series/implementations/decimal.rs
@@ -326,6 +326,14 @@ impl SeriesTrait for SeriesWrap<DecimalChunked> {
         self.0.get_any_value_unchecked(index)
     }
 
+    fn top_k(&self, k: usize, descending: bool) -> PolarsResult<Series> {
+        Ok(self
+            .0
+            .top_k(k, descending)
+            .into_decimal_unchecked(self.0.precision(), self.0.scale())
+            .into_series())
+    }
+
     fn sort_with(&self, options: SortOptions) -> PolarsResult<Series> {
         Ok(self
             .0
