+++ crates/polars-core/src/series/implementations/duration.rs
@@ -410,6 +410,14 @@ impl SeriesTrait for SeriesWrap<DurationChunked> {
         self.0.get_any_value_unchecked(index)
     }
 
+    fn top_k(&self, k: usize, descending: bool) -> PolarsResult<Series> {
+        Ok(self
+            .0
+            .top_k(k, descending)
+            .into_duration(self.0.time_unit())
+            .into_series())
+    }
+
     fn sort_with(&self, options: SortOptions) -> PolarsResult<Series> {
         Ok(self
             .0
