+++ crates/polars-core/src/series/implementations/mod.rs
@@ -329,6 +329,10 @@ macro_rules! impl_dyn_series {
                 self.0.get_any_value_unchecked(index)
             }
 
+            fn top_k(&self, k: usize, descending: bool) -> PolarsResult<Series> {
+                Ok(self.0.top_k(k, descending).into_series())
+            }
+
             fn sort_with(&self, options: SortOptions) -> PolarsResult<Series> {
                 Ok(ChunkSort::sort_with(&self.0, options).into_series())
             }
